add menu to add-form.component

+++ src/app/add-form/add-form.component.ts
@@ -11,10 +11,12 @@ export class AddFormComponent implements OnInit {
   constructor() { }
 
   item = new FormControl('');
+  search = new FormControl('');
   amount = new FormControl('');
   key = 'items' ;
 
   itemArray = [];
+  controls = '';
 
   addItem() {
     if (this.item.value !== '' && this.amount.value !== '') {
@@ -23,6 +25,7 @@ export class AddFormComponent implements OnInit {
         amount: this.amount.value,
       });
       this.saveItems();
+      this.clearForm();
     }
   }
 
@@ -32,14 +35,32 @@ export class AddFormComponent implements OnInit {
   }
 
   clearItems() {
-    localStorage.clear();
-    this.itemArray = [];
+    if (confirm ('Delete all items from this list?')) {
+      localStorage.clear();
+      this.itemArray = [];
+    }
+  }
+
+  deleteItem (i) {
+    this.itemArray.splice(i , 1);
+    this.saveItems();
+  }
+
+  clearForm () {
+    this.amount.setValue('');
+    this.item.setValue('');
+  }
+
+  controlSwitch(val) {
+    this.controls = val;
+  }
+
+  onSearch () {
   }
 
   ngOnInit() {
     if (JSON.parse(localStorage.getItem(this.key)) !== null) {
       this.itemArray = JSON.parse(localStorage.getItem(this.key));
+      }
     }
   }
-
-}
